Move profile patching out of the person module
service/api/person/__init__.py had grown to 2,286 lines, of which the
profile get/delete/patch endpoints and everything they dispatch through were
a self-contained 620. They move to person/profileinfo, together with the two
helpers nothing else called (`_has_gold` and `_str_value`).

Two smaller groups move with them because both the profile endpoints and the
handlers left behind need them: image processing and object-store upload to
person/images, and the rude-text and banned-photo checks to person/rudecheck,
whose entry point loses its underscore now that it is called across modules.
service/api/__init__.py routes the three endpoints to their new home, and the
26 imports the move left unused are dropped.

Nothing is rewritten: every block in the new modules appears verbatim in the
old file, and so does every block left in it.
